app/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-debug-toolbar==5.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==21.2.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.6
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
